+++ src/main/resources/db/changelog/changeset/10.create-orders-table.sql
@@ -2,8 +2,8 @@
 --changeset Alexey Savchic:orders-table
 create table orders
 (
-id serial primary key,
-user_id integer references users (id) on delete cascade not null,
+id bigserial primary key,
+user_id bigint references users (id) on delete cascade not null,
 order_date timestamp without time zone not null,
-summary_price integer check (summary_price>=0) not null
+summary_price decimal(7,2) check (summary_price>=0) not null
 );
