+++ src/main/resources/db/changelog/changeset/2.create-role-table.sql
@@ -2,6 +2,6 @@
 --changeset Alexey Savchic:role-table
 create table role
 (
-user_id integer references users (id) on delete cascade not null,
+user_id bigint references users (id) on delete cascade not null,
 role varchar(20) not null
 );
